WIP: v4l2_codec2: Use vendor_available for libc2plugin_store
Currently when using the v4l2_codec2 HAL as documented, it will
fail to function and one will see the following error in the
logcat:
03-11 05:25:09.121   568  2059 D CCodecBufferChannel: [c2.v4l2.avc.decoder#295] Created input block pool with allocatorID 16 => poolID 19 - OK (0)
03-11 05:25:09.122   568  2059 D C2PlatformStorePluginLoader: Failed to load library: libc2plugin_store.so (dlopen failed: library "libc2plugin_store.so" not found)
03-11 05:25:09.122   568  2059 D C2PlatformStorePluginLoader: Handle or CreateAllocator symbol is null

It seems this is due to the C2PlatformStorePluginLoader being
system library which tries to load and call out to
libc2plugin_store, which isn't allowed (vendor can call to
system, but system can't call to vendor).

This patch sets libc2plugin_store library to be vendor_available
which sets it in the system image.

Apparently this isn't the right solution, but without it
the libraries will not load.

It seems the libc2plugin_store loader logic may not be valid
post-Treble?
https://cs.android.com/android/platform/superproject/+/master:frameworks/av/media/codec2/vndk/C2PlatformStorePluginLoader.cpp;l=38

We need to understand what a proper solution would be, or what
vendors use for this.

Bug: 170199771
Test: none
Signed-off-by: Dmitry Shmidt <dimitrysh@google.com>
[jstultz: Added some detail to the commit message, narrowed
 change to just libc2plugin_store]
Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I7e4486ad95fb32a75ae9cec075672de0c7de2586